DCMAW-9802: Fix sonar coverage
