+++ sonar-project.properties
@@ -6,3 +6,7 @@ sonar.sources=backend-api/src
 sonar.javascript.lcov.reportPaths=backend-api/coverage/lcov.info
 sonar.language=ts
 sonar.exclusions=**/*.test.ts,**/*.test.js,**/testUtils/**
+
+# This is tested by E2E tests - it is not possible to test via unit tests.
+# Changes to this file causes quality gates to fail
+sonar.cpd.exclusions=**/functions/handlerDependencies.ts
